requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
google-genai>=1.0.0
pydantic>=2.9.0
python-dotenv>=1.0.0
jinja2>=3.1.0
networkx>=3.4
